+++ CHANGELOG.md
@@ -21,6 +21,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 --- -->
 
+## [Unreleased]
+
+### Added
+
+- Add removeFromShoppingList handler and update item references
+
+---
+
 ## [0.2.5] - 2026-03-03
 
 ### Added